requirements.txt
Django==5.1.3
djangorestframework==3.15.2
psycopg2==2.9.9
drf-spectacular==0.27.2
requests==2.32.3
django-redis==5.4.0
